backend/voshan/ml-service/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-compress>=1.14
python-socketio>=5.10.0
pillow>=10.0.0
pyyaml>=6.0
gunicorn>=21.2.0

